+++ Manifest.txt
@@ -7,17 +7,15 @@ README.rdoc
 Rakefile
 docs/bsdl.txt
 docs/ruby.txt
-lib/archive-tar-minitar.rb
-lib/archive/tar/minitar.rb
-lib/archive/tar/minitar/input.rb
-lib/archive/tar/minitar/output.rb
-lib/archive/tar/minitar/posix_header.rb
-lib/archive/tar/minitar/reader.rb
-lib/archive/tar/minitar/writer.rb
 lib/minitar.rb
-support/hoe/deprecated_gem.rb
+lib/minitar/input.rb
+lib/minitar/output.rb
+lib/minitar/posix_header.rb
+lib/minitar/reader.rb
+lib/minitar/writer.rb
 test/minitest_helper.rb
 test/support/tar_test_helpers.rb
+test/test_issue_46.rb
 test/test_tar_header.rb
 test/test_tar_input.rb
 test/test_tar_output.rb
