--- lib/archive-tar-minitar.rb
@@ -1,3 +1,0 @@
-# coding: utf-8
-
-require "archive/tar/minitar"
